cleanup for pr #48
