+++ src/keepalive.rs
@@ -129,7 +129,6 @@ pub async fn watch_dummy(client: Arc<SplinterClient>, dummy_conn: Arc<SplinterSe
                 break debug!("dummy conn {} no longer alive", dummy_conn.server.id);
             }
             let mut lock = dummy_conn.reader.lock().await;
-            //let packet_read_result = smol::future::or(lock.read_raw_packet_async::<RawPacketLatest>().await
             let raw_packet = match lock.read_raw_packet_async::<RawPacketLatest>().await {
                 Ok(Some(packet)) => packet,
                 Ok(None) => {
@@ -155,7 +154,6 @@ pub async fn watch_dummy(client: Arc<SplinterClient>, dummy_conn: Arc<SplinterSe
                     Ok(packet) => match packet {
                         PacketLatest::PlayServerKeepAlive(body) => {
                             let mut writer = dummy_conn.writer.lock().await;
-                            //debug!("{}-{} got keep alive", &client.name, dummy_conn.server.id);
                             if let Err(e) = (*writer).write_packet_async(PacketLatest::PlayClientKeepAlive(PlayClientKeepAliveSpec {
                                 id: body.id
                             })).await {
@@ -186,20 +184,18 @@ pub async fn watch_dummy(client: Arc<SplinterClient>, dummy_conn: Arc<SplinterSe
                             }
                             // if the position the server wants us to go to is farther than where
                             // we actually should be, then send a position set to the plugin
-                            // get dist
+
+                            // as a note here, this only handles when the provided teleportation
+                            // request has an absolute position. TODO: relative position
                             if body.flags.0 == 0 {
-                                debug!("lucky us.. got an absolute position request");
                                 let tpos = body.location.position;
                                 let ppos = &client.position.load().as_ref().unwrap();
                                 const MAX_DIST: f64 = 15.;
-                                debug!("comparison (t: {:?}, p: {:?})", &tpos, &ppos);
                                 if (tpos.x - ppos.x).abs() > MAX_DIST || (tpos.y - ppos.y).abs() > MAX_DIST || (tpos.z - ppos.z).abs() > MAX_DIST {
-                                    debug!("too far!");
                                     if let Err(e) = send_position_set(writer, ppos.x, ppos.y, ppos.z).await {
                                         dummy_conn.alive.store(false, Ordering::Relaxed);
                                         break error!("Failed to send position set to dummy {}-{}: {:?}", &client.name, dummy_conn.server.id, e);
                                     }
-                                    debug!("sent position set");
                                 }
                             }
                         },
@@ -214,25 +210,13 @@ pub async fn watch_dummy(client: Arc<SplinterClient>, dummy_conn: Arc<SplinterSe
                     }
                 }
             }
-            else {
-                debug!("{}-{}: {:?}", &client.name, dummy_conn.server.id, packet_kind);
-            }
             if has_eids(lazy_packet.kind()) {
                 if let Ok(packet) = lazy_packet.packet() {
                     let map = &mut *client.proxy.mapping.lock().await;
                     pass_through = pass_through || SplinterMappingResult::Client == map_eid(&*client, map, packet, &PacketDirection::ClientBound, &dummy_conn.server);
                 }
             }
-            /*if has_uuids(lazy_packet.kind()) {
-                if let Ok(packet) = lazy_packet.packet() {
-                    let map = &mut *client.proxy.mapping.lock().await;
-                    // yes, this is &&, not ||. if map uuid says no, we override anything map eid
-                    // said
-                    pass_through = pass_through && SplinterMappingResult::Client == map_uuid(map, packet, &PacketDirection::ClientBound, &dummy_conn.server);
-                }
-            }*/
             if pass_through {
-                //debug!("passing through a {:?}", packet_kind);
                 if let Err(e) = send_packet(&client, &PacketDestination::Client, lazy_packet)
                     .await
                     .with_context(|| {
