+++ src/protocol/v_cur/chunk.rs
@@ -2,9 +2,16 @@ use super::{
     PacketDestination,
     RelayPass,
 };
-use crate::current::{
-    PacketLatest,
-    PacketLatestKind,
+use crate::{
+    client::{
+        ChunkLoadData,
+        SplinterClient,
+    },
+    current::{
+        PacketLatest,
+        PacketLatestKind,
+    },
+    server::SplinterServerConnection,
 };
 
 inventory::submit! {
@@ -35,10 +42,64 @@ inventory::submit! {
                 if !pass_through {
                     *destination = PacketDestination::None;
                 }
-                /*else {
-                    debug!("active through {:?}", lazy_packet.kind());
-                }*/
             }
         }
     }))
+}
+
+impl SplinterServerConnection {
+    /// Returns whether we pass the packet on
+    pub async fn update_chunk(
+        &self,
+        client: &SplinterClient,
+        is_chunkdata: bool,
+        chunk: (i32, i32),
+    ) -> bool {
+        let newly_added_to_self = self.known_chunks.lock().await.insert(chunk);
+        let client_known_chunks = &mut *client.known_chunks.lock().await;
+        if let Some(load_data) = client_known_chunks.get_mut(&chunk) {
+            if newly_added_to_self {
+                load_data.refcount += 1;
+            }
+            if is_chunkdata {
+                if !load_data.received_chunkdata {
+                    load_data.received_chunkdata = true;
+                    true
+                } else {
+                    false
+                }
+            } else {
+                if !load_data.received_updatelight {
+                    load_data.received_updatelight = true;
+                    true
+                } else {
+                    false
+                }
+            }
+        } else {
+            client_known_chunks.insert(
+                chunk,
+                ChunkLoadData {
+                    received_chunkdata: is_chunkdata,
+                    received_updatelight: !is_chunkdata,
+                    refcount: 1,
+                },
+            );
+            true
+        }
+    }
+    pub async fn remove_chunk(&self, client: &SplinterClient, chunk: (i32, i32)) -> bool {
+        if self.known_chunks.lock().await.remove(&chunk) {
+            let client_known_chunks = &mut *client.known_chunks.lock().await;
+            if let Some(load_data) = client_known_chunks.get_mut(&chunk) {
+                if load_data.refcount > 1 {
+                    load_data.refcount -= 1;
+                } else {
+                    client_known_chunks.remove(&chunk);
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
 }
